Fix STATIC_INLINE for gcc -std=c89
According to [1], the __GNUC__ macro is always defined for a gcc-compilant
compiler, whether it is in strict C standard mode or in GNU mode. This
causes the STATIC_INLINE macro to always expand to "static inline" on
gcc/clang, even in C89 mode where "inline" is not defined.

Unfortunately, there does not seem to be a gcc macro that checks for GNU
mode, so this is fixed by removing the check on __GNUC__.

[1]: https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html